Split macros.js into separate files under macros/
