+++ src/apps/events/macros/event-edit-form.js
@@ -1,11 +1,6 @@
-const { assign } = require('lodash')
-
-const metadataRepo = require('../../lib/metadata')
-const { transformObjectToOption } = require('../transformers')
-const { globalFields } = require('../macros')
-const { collectionFilterLabels } = require('./labels')
-
-const currentYear = (new Date()).getFullYear()
+const metadataRepo = require('../../../lib/metadata')
+const { globalFields } = require('../../macros')
+const { transformObjectToOption } = require('../../transformers')
 
 const eventFormConfig = ({ advisers }) => {
   return {
@@ -86,7 +81,7 @@ const eventFormConfig = ({ advisers }) => {
         label: 'Postcode',
         class: 'u-js-hidden',
       },
-      assign({}, globalFields.countries, {
+      Object.assign({}, globalFields.countries, {
         name: 'address_country',
       }),
       {
@@ -179,70 +174,4 @@ const eventFormConfig = ({ advisers }) => {
   }
 }
 
-const eventFiltersFields = [
-  {
-    macroName: 'TextField',
-    name: 'name',
-    hint: 'At least three characters',
-  },
-  assign({}, globalFields.eventTypes, {
-    name: 'event_type',
-    initialOption: 'All types',
-  }),
-  assign({}, globalFields.countries, {
-    name: 'address_country',
-    initialOption: 'All countries',
-  }),
-  assign({}, globalFields.ukRegions, {
-    name: 'uk_region',
-    initialOption: 'All UK regions',
-  }),
-  {
-    macroName: 'TextField',
-    name: 'start_date',
-    hint: 'YYYY-MM-DD',
-    placeholder: `e.g. ${currentYear}-07-18`,
-  },
-  {
-    macroName: 'TextField',
-    name: 'end_date',
-    hint: 'YYYY-MM-DD',
-    placeholder: `e.g. ${currentYear}-07-21`,
-  },
-].map(filter => {
-  return assign(filter, {
-    label: collectionFilterLabels.edit[filter.name],
-    modifier: ['smaller', 'light'],
-  })
-})
-
-const eventSortForm = {
-  method: 'get',
-  class: 'c-collection__sort-form js-AutoSubmit',
-  hideFormActions: true,
-  hiddenFields: { custom: true },
-  children: [
-    {
-      macroName: 'MultipleChoiceField',
-      label: 'Sort by',
-      name: 'sortby',
-      modifier: ['small', 'inline', 'light'],
-      options: [
-        { value: 'name:asc', label: 'Event name: A-Z' },
-        { value: 'name:desc', label: 'Event name: Z-A' },
-        { value: 'modified_on:desc', label: 'Recently updated' },
-        { value: 'modified_on:asc', label: 'Least recently updated' },
-        { value: 'start_date:asc', label: 'Earliest start date' },
-        { value: 'start_date:desc', label: 'Latest start date' },
-        { value: 'end_date:asc', label: 'Earliest end date' },
-        { value: 'end_date:desc', label: 'Latest end date' },
-      ],
-    },
-  ],
-}
-
-module.exports = {
-  eventFormConfig,
-  eventFiltersFields,
-  eventSortForm,
-}
+module.exports = eventFormConfig
