+++ src/apps/events/macros/index.js
@@ -1,0 +1,9 @@
+const eventFormConfig = require('./event-edit-form')
+const eventFiltersFields = require('./event-filters-fields')
+const eventSortForm = require('./event-sort-form')
+
+module.exports = {
+  eventFormConfig,
+  eventFiltersFields,
+  eventSortForm,
+}
